Move shared webpack config under @wordpress/scripts package
