+++ packages/scripts/utils/string.js
