--- packages/webpack-config/CHANGELOG.md
@@ -1,3 +1,0 @@
-## 1.0.0 (Unreleased)
-
-- Initial release.
