--- packages/webpack-config/index.js
@@ -1,10 +1,0 @@
-/**
- * Internal dependencies
- */
-const { camelCaseDash } = require( './utils' );
-const config = require( './config' );
-
-module.exports = {
-	camelCaseDash,
-	config,
-};
