--- packages/webpack-config/package.json
@@ -1,34 +1,0 @@
-{
-	"name": "@wordpress/webpack-config",
-	"version": "1.0.0-alpha.0",
-	"description": "Webpack config for WordPress development.",
-	"author": "The WordPress Contributors",
-	"license": "GPL-2.0-or-later",
-	"keywords": [
-		"wordpress",
-		"webpack",
-		"config"
-	],
-	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/webpack-config/README.md",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/WordPress/gutenberg.git"
-	},
-	"bugs": {
-		"url": "https://github.com/WordPress/gutenberg/issues"
-	},
-	"engines": {
-		"node": ">=8",
-		"npm": ">=6.0.0"
-	},
-	"main": "index.js",
-	"dependencies": {
-		"babel-loader": "^8.0.5",
-		"source-map-loader": "^0.2.4",
-		"webpack-bundle-analyzer": "^3.0.3",
-		"webpack-livereload-plugin": "^2.2.0"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
-}
